Update rental to 0.5.6
Your crate depends on an old version of `rental`, which relies on
a backwards-compatibiltiy hack in rustc. In future versions of rustc,
versions of `rental` below `0.5.6` will now compiling, breaking your
crate.

This PR updates your crate to `rental` 0.5.6, which ensures that it will
continue to compile with both older and newer version of rustc.

For more information, see https://github.com/rust-lang/rust/issues/83125